+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Changed the returned `SignatureVerificationError.Status` when trying to verify a message with no embedded signature. It used to return `constants.SIGNATURE_NO_VERIFIER` and now returns `constants.SIGNATURE_NOT_SIGNED`.
+This change impacts : 
+	- `func (sk *SessionKey) DecryptAndVerify(...)`
+	- `func (msg *PlainMessageReader) VerifySignature(...)`
+	- `func (keyRing *KeyRing) Decrypt(...)`
+
+### Added 
+- Helper to access the SignatureVerificationError explicitly when decrypting streams  in mobile apps: 
+	```go
+	func VerifySignatureExplicit(
+		reader *crypto.PlainMessageReader,
+	) (signatureVerificationError *crypto.SignatureVerificationError, err error)
+	```
+
 ## [2.2.0] 2021-06-30
 ### Added
 - Streaming API: