+++ CHANGELOG.md
@@ -4,15 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
-
+## [2.2.1] 2021-07-27
 ### Changed
-
 - Changed the returned `SignatureVerificationError.Status` when trying to verify a message with no embedded signature. It used to return `constants.SIGNATURE_NO_VERIFIER` and now returns `constants.SIGNATURE_NOT_SIGNED`.
 This change impacts : 
 	- `func (sk *SessionKey) DecryptAndVerify(...)`
 	- `func (msg *PlainMessageReader) VerifySignature(...)`
 	- `func (keyRing *KeyRing) Decrypt(...)`
+- Improved error messages for failures in password protected message decryption
 
 ### Added 
 - Helper to access the SignatureVerificationError explicitly when decrypting streams  in mobile apps: